requirements.txt
numpy==1.19.2
matplotlib==3.3.2
stylecloud==0.5.1
wordcloud==1.8.1
streamlit==0.79.0
Pillow==8.1.2
